Refactor to use create_io for seek & iter_both
store_read was left as is for now, because it deals with only one file
at a time and can't use create_io easily.
