--- data/ncbi_2_unpaired.fastq
@@ -1,4 +1,0 @@
-@SRR3380692.5.2 3 length=101
-CCAACGTGGCAAAACCCCGTCTCTACTAAAAATACAAAAAAATTAGCCAGACGTGGTGACAGGTGCCTGTAGTCCCAGCTACTCAGGAGGTTGAGATAGGA
-+
-CCCFFFFFHHHHHIJJJJJJJJJJJJJIIIIJJJJJIIIJJJJJJIJJJJJJHHFFEFEEEEDCDDDDDCDDEDDDDDDDDDDDDDCDDD<ACDDDDDDDD
