+++ data/ncbi_unpaired.fastq
@@ -1,3 +1,7 @@
+@SRR3380692.5.2 3 length=101
+CCAACGTGGCAAAACCCCGTCTCTACTAAAAATACAAAAAAATTAGCCAGACGTGGTGACAGGTGCCTGTAGTCCCAGCTACTCAGGAGGTTGAGATAGGA
++
+CCCFFFFFHHHHHIJJJJJJJJJJJJJIIIIJJJJJIIIJJJJJJIJJJJJJHHFFEFEEEEDCDDDDDCDDEDDDDDDDDDDDDDCDDD<ACDDDDDDDD
 @SRR3380692.9.1 3 length=101
 CTCACNGCAACCTCCGTCTCCTGGGTTCAAGCGATTTTCCTATCTCAACCTCCTGAGTAGCTGGGACTACAGGCACCTGTCACCACGTCTGGCTAATTTTT
 +
